pyproject.toml
[project]
name = "advocacy-bot"
version = "0.1.0"
description = "Discord bot for San Diego City Council agenda tracking"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3,<3.0",
    "httpx>=0.27,<1.0",
    "beautifulsoup4>=4.12,<5.0",
    "aiosqlite>=0.19,<1.0",
    "python-dotenv>=1.0,<2.0",
]

[project.scripts]
advocacy-bot = "advocacy_bot.bot:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
